Separate code to reset version to 0

+++ counter.py
@@ -1,28 +1,49 @@
 from datetime import datetime
 import json
 import re
+import sys
 
+if sys.argv[1] == 'startup':
+    with open('counter.json', 'r+') as f:
+        data = json.load(f)
 
-with open('counter.json', 'r+') as f:
-    data = json.load(f)
-    if data['version'] != datetime.today().strftime('%Y.%m%d'):
-        data['version'] = datetime.today().strftime('%Y.%m%d')
-        data['build'] = 0
-    else:
+        if data['version'] != datetime.today().strftime('%Y.%m%d'):
+            data['version'] = datetime.today().strftime('%Y.%m%d')
+            data['build'] = 0
+        new_version = f'{data["version"]}.{data["build"]}'
+
+        g = open('constants.py', 'r')
+        lines = g.readlines()
+        g.close()
+
+        with open('constants.py', 'w') as g:
+            version = re.findall(r'\d{1,4}\.\d{1,4}\.\d{1,3}', lines[10])
+            if version:
+                lines[10] = lines[10].replace(version[0], new_version)
+            for line in lines:
+                g.write(line)
+
+        f.seek(0)
+        json.dump(data, f, indent=2)
+        f.truncate()
+
+if sys.argv[1] == 'update':
+    with open('counter.json', 'r+') as f:
+        data = json.load(f)
         data['build'] += 1
-    new_version = f'{data["version"]}.{data["build"]}'
+        new_version = f'{data["version"]}.{data["build"]}'
 
-    g = open('constants.py', 'r')
-    lines = g.readlines()
-    g.close()
+        g = open('constants.py', 'r')
+        lines = g.readlines()
+        g.close()
 
-    with open('constants.py', 'w') as g:
-        version = re.findall(r'\d{1,4}\.\d{1,4}\.\d{1,3}', lines[10])
-        if version:
-            lines[10] = lines[10].replace(version[0], new_version)
-        for line in lines:
-            g.write(line)
+        with open('constants.py', 'w') as g:
+            version = re.findall(r'\d{1,4}\.\d{1,4}\.\d{1,3}', lines[10])
+            if version:
+                lines[10] = lines[10].replace(version[0], new_version)
+            for line in lines:
+                g.write(line)
 
-    f.seek(0)
-    json.dump(data, f, indent=2)
-    f.truncate()
+        f.seek(0)
+        json.dump(data, f, indent=2)
+        f.truncate()
